TPI-276 Implementing sending of notifications when an account is added, deleted or updated, and when an account contact is added, deleted or updated.

+++ RLS.PortfolioGeneration.FrontendBackend/Controllers/AccountContactController.cs
@@ -4,7 +4,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Options;
 using RLS.PortfolioGeneration.FrontendBackend.Dtos;
+using RLS.PortfolioGeneration.FrontendBackend.Services;
 using RLS.PortfolioGeneration.Persistence.Model;
 using RLS.PortfolioGeneration.Persistence.Model.Clients;
 
@@ -14,10 +16,13 @@ namespace RLS.PortfolioGeneration.FrontendBackend.Controllers
     public class AccountContactController : Controller
     {
         private readonly ModelDbContext _dbContext;
+        private AccountReportingService _reportingService;
 
-        public AccountContactController(ModelDbContext dbContext)
+        public AccountContactController(ModelDbContext dbContext, IOptions<HierarchyServiceConfiguration> configuration)
         {
             _dbContext = dbContext;
+
+            this._reportingService = new AccountReportingService(configuration.Value);
         }
 
         [HttpGet]
@@ -56,6 +61,7 @@ namespace RLS.PortfolioGeneration.FrontendBackend.Controllers
 
             await _dbContext.Add(contact);
 
+            await _reportingService.ReportAccountChanged(contactDto.AccountId.ToString());
             return Created(new Uri($"/api/Contact/{contactId}", UriKind.Relative), new { id = contactId });
         }
 
@@ -65,13 +71,20 @@ namespace RLS.PortfolioGeneration.FrontendBackend.Controllers
             var contact = Mapper.Map<AccountContact>(contactDto);
             contact.Id = id;
 
+            await _reportingService.ReportAccountChanged(contactDto.AccountId.ToString());
             await _dbContext.Update(contact);
         }
 
         [HttpDelete("{id}")]
         public async Task Delete(Guid id)
         {
-            await _dbContext.DeleteContact(id);
+            var contact = await _dbContext.RetrieveContactById(id);
+            if (contact != null)
+            {
+                await _reportingService.ReportAccountChanged(contact.AccountId.ToString());
+
+                await _dbContext.DeleteContact(id);
+            }
         }
     }
 }
